requirements.txt
Django==3.0.3
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
-e git+https://github.com/rodrigoegimenez/spectranalyzer.git@spectranalyzer2#egg=SpectrAnalyzer
